requirements.txt.txt
Flask==2.3.2
mysql-connector-python==8.0.33
ultralytics==8.0.196
Pillow==9.5.0
numpy==1.24.3
requests==2.28.1
